Refactor prior to introduction of new utilities
